app/requirements.txt
flask==3.1.1
pytest==8.3.5
requests==2.32.3
gunicorn==23.0.0
python-dotenv==1.1.0
prometheus-flask-exporter==0.23.0
